requirements.txt
python-telegram-bot>=21.0
google-genai>=1.0.0
numpy
pytest>=8.0
